feat: create backend environment for event feature
As I am implementing new events feature to the application, backend is
ofcourse needed as well

As One may expect there will be quite a few events, I decided to create
a paginated api to limit the number of entries a user can receive at
once, which I also believe will be better in term of performace. I am
though keeping the possibility of requesting all of the events at once
opened in case its needed im the future

When it comes to posting new events, it is fairly simple, I am
validating the request body just as I did with users and adding a check
to see if the events does not already exist throughthe events name.
I can imagine having two same events in the API would only be confusing
for both the user and me as there would be duplicates which we would
have to remove on the frontend side, it would be painfully complicated
for no reason.

updating is farily easy, I decided to go with findOneAndupdate method
looking for the events name before the change as it is stored and
accessible and if the events is found, it gets updated.

Deleting the event is very basic as there is nothing of high importance
needed